no message
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@63 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/misc.cpp
@@ -156,7 +156,7 @@ static int ilog2(const float x)
 }
 
 
-long round_up(long val, int multiple)
+uintptr_t round_up(uintptr_t val, uintptr_t multiple)
 {
 	val += multiple-1;
 	val -= val % multiple;
